+++ CHANGELOG.md
@@ -5,6 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Option for treating not-nullable fields as "required" in their parent type
+- Option for treating not-nullable methods as "required" in their parent type
+- Indicate a string's "format" to be "email" if `@Email` is present
+- Option for returning "idn-email" instead of "email" as "format" if `@Email` is present
+- Indicate a string's "pattern" according to regular expressions on `@Pattern` or `@Email` (ignoring specified flags)
+- Option for enabling the inclusion of "pattern" expressions (they are excluded by default)
 
 ## [3.0.0] – 2019-06-10
 ### Added